# contracts/openapi.yaml
components:
  schemas:
...
    ActivityItemResponse:
      description: Activity item response.
      properties:
        action:
          description: Action performed.
          example: create
          type: string
        amount:
          description: Amount (if applicable).
          type:
          - string
          - 'null'
        currency:
          description: Currency (if applicable).
          type:
          - string
          - 'null'
        description:
          description: Description.
          example: New transaction created for 120.00 USD
          type: string
        entity_id:
          description: Entity ID.
          format: uuid
          type: string
        entity_type:
          description: Entity type.
          example: transaction
          type: string
        id:
          description: Activity ID.
          format: uuid
          type: string
        timestamp:
          description: Timestamp.
          type: string
        type:
          description: Activity type.
          example: transaction_create
          type: string
        user:
          $ref: '#/components/schemas/UserInfo'
          description: User info.
      required:
      - id
      - type
      - action
      - entity_type
      - entity_id
      - description
      - user
      - timestamp
      type: object
...
    BudgetLineItemResponse:
      description: Budget line item response.
      properties:
        account_code:
          description: Account code.
          example: '6001'
          type: string
        account_id:
          description: Account ID.
          format: uuid
          type: string
        account_name:
          description: Account name.
          example: Office Expenses
          type: string
        actual:
          description: Actual amount.
          type: string
        budgeted:
          description: Budgeted amount.
          type: string
        variance:
          description: Variance.
          type: string
        variance_percent:
          description: Variance percentage.
          format: double
          type: number
      required:
      - account_id
      - account_code
      - account_name
      - budgeted
      - actual
      - variance
      - variance_percent
      type: object
...
    BudgetSummary:
      description: Budget summary.
      properties:
        total_actual:
          description: Total actual.
          type: string
        total_budgeted:
          description: Total budgeted.
          type: string
        variance:
          description: Variance (budgeted - actual).
          type: string
        variance_percent:
          description: Variance percentage.
          format: double
          type: number
      required:
      - total_budgeted
      - total_actual
      - variance
      - variance_percent
      type: object
    BudgetVsActualResponse:
      description: Response for budget vs actual data.
      properties:
        budget_id:
          description: Budget ID.
          format: uuid
          type:
          - string
          - 'null'
        budget_name:
          description: Budget name.
          example: Annual Budget 2023
          type:
          - string
          - 'null'
        line_items:
          description: Line items.
          items:
            $ref: '#/components/schemas/BudgetLineItemResponse'
          type: array
        summary:
          $ref: '#/components/schemas/BudgetSummary'
          description: Summary.
      required:
      - summary
      - line_items
      type: object
...
    CashFlowDataPoint:
      description: Cash flow data point.
      properties:
        inflow:
          description: Total inflow.
          type: string
        month:
          description: Month label (e.g., "Jan").
          example: Jan
          type: string
        net:
          description: Net cash flow.
          type: string
        outflow:
          description: Total outflow.
          type: string
        period_name:
          description: Period name (e.g., "2026-01").
          example: 2026-01
          type: string
      required:
      - month
      - period_name
      - inflow
      - outflow
      - net
      type: object
    CashFlowResponse:
      description: Response for cash flow data.
      properties:
        data:
          description: Cash flow data points.
          items:
            $ref: '#/components/schemas/CashFlowDataPoint'
          type: array
      required:
      - data
      type: object
...
    PaginationInfo:
      description: Pagination info.
      properties:
        has_more:
          description: Has more results.
          type: boolean
        limit:
          description: Limit.
          format: int64
          minimum: 0
          type: integer
        next_cursor:
          description: Next cursor.
          type:
          - string
          - 'null'
      required:
      - limit
      - has_more
      type: object
...
    RecentActivityResponse:
      description: Response for recent activity.
      properties:
        activities:
          description: Activity items.
          items:
            $ref: '#/components/schemas/ActivityItemResponse'
          type: array
        pagination:
          $ref: '#/components/schemas/PaginationInfo'
          description: Pagination info.
      required:
      - activities
      - pagination
      type: object
...
    UserInfo:
      description: User info.
      properties:
        full_name:
          description: Full name.
          example: John Doe
          type: string
        id:
          description: User ID.
          format: uuid
          type: string
      required:
      - id
      - full_name
      type: object
...
  securitySchemes:
    bearerAuth:
      bearerFormat: JWT
      scheme: bearer
      type: http
info:
  description: HTTP API layer with Axum routes and middleware
  license:
    name: ''
  title: zeltra-api
  version: 0.1.0
openapi: 3.1.0
paths:
...
  /organizations/{org_id}/budgets/{budget_id}/vs-actual:
    get:
      operationId: get_budget_vs_actual
      parameters:
      - description: Organization ID
        in: path
        name: org_id
        required: true
        schema:
          format: uuid
          type: string
      - description: Budget ID
        in: path
        name: budget_id
        required: true
        schema:
          format: uuid
          type: string
      - description: Filter by fiscal period ID.
        in: path
        name: fiscal_period_id
        required: true
        schema:
          format: uuid
          type:
          - string
          - 'null'
      - description: Filter by dimension value IDs (comma-separated).
        in: path
        name: dimensions
        required: true
        schema:
          type:
          - string
          - 'null'
      responses:
        '200':
          description: Budget vs actual comparison
        '403':
          description: Forbidden
        '404':
          description: Budget not found
      security:
      - bearerAuth: []
      summary: GET `/organizations/{org_id}/budgets/{budget_id}/vs-actual` - Get budget vs actual comparison.
      tags:
      - Budgets
  /organizations/{org_id}/dashboard/budget-vs-actual:
    get:
      description: 'Requirement 4.7: Budget vs actual summary'
      operationId: get_budget_vs_actual
      parameters:
      - description: Organization ID
        in: path
        name: org_id
        required: true
        schema:
          format: uuid
          type: string
      - description: Budget ID (optional, uses first active budget if not provided).
        in: path
        name: budget_id
        required: true
        schema:
          format: uuid
          type:
          - string
          - 'null'
      responses:
        '200':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/BudgetVsActualResponse'
          description: Budget vs actual performance data
        '403':
          description: Forbidden
      security:
      - bearerAuth: []
      summary: GET /organizations/{org_id}/dashboard/budget-vs-actual
      tags:
      - Dashboard
  /organizations/{org_id}/dashboard/cash-flow:
    get:
      description: 'Requirement 4.5: Cash flow chart data'
      operationId: get_cash_flow
      parameters:
      - description: Organization ID
        in: path
        name: org_id
        required: true
        schema:
          format: uuid
          type: string
      - description: Number of months to include.
        example: 6
        in: path
        name: months
        required: true
        schema:
          format: int32
          minimum: 0
          type:
          - integer
          - 'null'
      - description: Fiscal period ID.
        in: path
        name: period_id
        required: true
        schema:
          format: uuid
          type:
          - string
          - 'null'
      responses:
        '200':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CashFlowResponse'
          description: Cash flow historical data
        '403':
          description: Forbidden
      security:
      - bearerAuth: []
      summary: GET /organizations/{org_id}/dashboard/cash-flow
      tags:
      - Dashboard
...
  /organizations/{org_id}/dashboard/recent-activity:
    get:
      description: 'Requirement 17.1: Recent activity endpoint'
      operationId: get_recent_activity
      parameters:
      - description: Organization ID
        in: path
        name: org_id
        required: true
        schema:
          format: uuid
          type: string
      - description: Maximum number of items to return.
        example: 10
        in: path
        name: limit
        required: true
        schema:
          format: int64
          minimum: 0
          type:
          - integer
          - 'null'
      - description: Activity type filter.
        example: transaction
        in: path
        name: type
        required: true
        schema:
          type:
          - string
          - 'null'
      - description: Cursor for pagination.
        in: path
        name: cursor
        required: true
        schema:
          type:
          - string
          - 'null'
      responses:
        '200':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/RecentActivityResponse'
          description: Recent activity logs
        '403':
          description: Forbidden
      security:
      - bearerAuth: []
      summary: GET /organizations/{org_id}/dashboard/recent-activity
      tags:
      - Dashboard
...